deployment/streamlit/requirements.txt
opencv-python
numpy
Pillow==8.1.0
requests
streamlit
pybase64